Rework FireInTheHole and JavelinRockets
JavelinRockets will be placed into the same rank as Incinerator and
therefore needs a buff. Increased aim and range are related and are
similar to Incinerator's range and angle increases, so this commit moves
the racket scatter reduction from FireInTheHole to JavelinRockets.

FireInTheHole in turn receives a 25% blast radius increase to all
rockets.

	FireInTheHole  = +damage, +radius
	JavelinRockets = +aim, +range