Make smoke grenade activation non-turn-ending
Using smoke grenades is almost never the correct play because of the
opportunity cost of performing an offensive action. This buff makes
smoke grenades much more likely to be used in regular tactics.

Since this slightly overlaps with the functionality of RapidDeployment,
add mimic beacons to the list of "support" grenades affected by
RapidDeployment.